Request YAML & cope with JSON responses
When loading Swagger docs from a remote location the app now requests
YAML, then JSON, then anything else (in that order) so that
sites with the ability to send multiple formats return the correct
one.

It can now also cope with JSON being returned (which is automatically
converted into an object by .get).