OAuth2.0 flow in docker fixed
Copied oauth2-redirect.html from the swagger-ui
project into the dist folder and updated the
Dockerfile to include the file where it could
be served by nginx. This page is called by the
authroization server after the user is redirected
back. The page should add the access token provided
by the auth server to subsequent request.