fix(monaco-editor): avoid disposal of standalone services (#3517)
StandaloneServices is a singleton and can be initialized only once.
Subsequent initializations are noops. This has a side effect which
is inability to dispose of the services via StandaloneServices interface.
Individual services can be disposed of separately, but if one decides
to do that StandaloneServices will not able to initialize them again.