fix(spectral): Fix validation crashing the UI
Backport of changes introduced in #6. Not entirely sure what happens
here, this seems to be a timing issue.

Not giving an explicit `source` to the messages means they go into the
same pool as the default messages, this also means that they fall into
the same cleanup logic. The `newSpecErrBatch` function seems to mess with
that cleanup (our spectral messages clean up the "regular messages") and
the UI tries to render messages that no longer exist, crashing in the
process.

the `newThrownErrBatch` function doesn't seem to have these issues but
it requires a debounce for some reason or it won't render anything.